Move the "accumulation" module to the "trailer" module
