+++ lib/src/trailer/accumulation.rs
