+++ src/main.rs
@@ -30,9 +30,9 @@ use chrono::{FixedOffset, TimeZone};
 use clap::App;
 use git2::Commit;
 use libgitdit::issue::IssueRefType;
-use libgitdit::message::accumulation::{self, Accumulator};
-use libgitdit::trailer::{Trailer, PairsToTrailers};
 use libgitdit::message::LineIteratorExt;
+use libgitdit::trailer::accumulation::{self, Accumulator};
+use libgitdit::trailer::{Trailer, PairsToTrailers};
 use libgitdit::{Issue, Message, RemoteExt, RepositoryExt};
 use log::LogLevel;
 use std::fs::File;
